"Enhancements to the log function across multiple scripts"
This commit introduces improvements to the log function across multiple scripts. Specifically, the function now provides the option of adding a remark and a message. However, if these values are not provided, they are set to "default_remark" and "empty msg by log default" respectively. Moreover, a new logger.sh file was created in the commons folder, and subsequently sourced in the dd