Update command invocation syntax and add jq check
Changed the command invocation in ddns.sh and ip_cache.sh from /bin/bash to bash for portability. Also, added a function in ip_cache.sh to check if 'jq' command exists. The script will log the presence of the command and terminate if it does not exist, to ensure json handling capability.